chatbot Facebook/direct_service/admin_requirements.txt
google-api-python-client==2.176.0
google-auth==2.40.3
python-dotenv==1.1.1

